Added Idempotency and Implemented Choreo Saga

+++ src/main/java/com/example/LodgeBnB/models/Airbnb.java
@@ -21,7 +21,7 @@ public class Airbnb {
     private String description;
 
     @Column(nullable = false)
-    private Long pricePerNight;
+    private Double pricePerNight;
 
     private String location;
 }
